Eliminate createAddressHelper and fix Substrate login.
